@@ -15,74 +16,61 @@
     "photoswipe",
     "gallery"
   ],
-  "peerDependencies": {
-    "react": "^0.14 || ^15.0.0",
-    "react-dom": "^0.14 || ^15.0.0"
-  },
   "dependencies": {
     "classnames": "^2.2.3",
     "lodash.pick": "^4.2.1",
     "photoswipe": "^4.1.0"
   },
   "devDependencies": {
-    "autoprefixer": "^6.3.1",
-    "babel": "^5.8.34",
-    "babel-core": "^5.8.29",
-    "babel-eslint": "^4.1.3",
-    "babel-loader": "^5.3.2",
-    "babel-plugin-react-transform": "^1.1.1",
-    "bootstrap": "^3.3.6",
-    "camelcase": "^2.0.1",
-    "css-loader": "^0.23.0",
-    "cssnano": "^3.3.2",
-    "del": "^2.1.0",
-    "eslint": "^1.10.1",
-    "eslint-loader": "^1.2.0",
-    "eslint-plugin-react": "^3.15.0",
-    "express": "^4.13.3",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.8.5",
-    "gulp": "^3.9.0",
-    "gulp-autoprefixer": "^3.1.0",
-    "gulp-babel": "^5.2.1",
-    "gulp-concat": "^2.6.0",
-    "gulp-eslint": "^1.1.0",
-    "gulp-filter": "^3.0.1",
-    "gulp-minify-css": "^1.2.4",
-    "gulp-postcss": "^6.1.0",
+    "autoprefixer": "^7.0.1",
+    "babel-cli": "^6.8.0",
+    "babel-core": "^6.8.0",
+    "babel-eslint": "^7.1.1",
+    "babel-loader": "^7.0.0",
+    "babel-plugin-add-module-exports": "^0.2.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-object-rest-spread": "^6.23.0",
+    "babel-preset-env": "^1.2.2",
+    "babel-preset-react": "^6.24.1",
+    "bootstrap": "^3.3.7",
+    "camelcase": "^4.1.0",
+    "clean-webpack-plugin": "^0.1.9",
+    "cross-env": "^4.0.0",
+    "css-loader": "^0.28.0",
+    "cssnano": "^3.5.2",
+    "del": "^3.0.0",
+    "eslint": "^3.15.0",
+    "eslint-config-airbnb": "^15.0.1",
+    "eslint-loader": "^1.3.0",
+    "eslint-plugin-import": "^2.3.0",
+    "eslint-plugin-jsx-a11y": "^5.0.3",
+    "eslint-plugin-react": "^7.0.1",
+    "express": "^4.15.3",
+    "extract-text-webpack-plugin": "^2.1.0",
+    "file-loader": "^0.11.1",
+    "flux": "^3.1.2",
+    "globby": "^6.1.0",
+    "gulp": "^3.9.1",
+    "gulp-babel": "^6.1.2",
+    "gulp-concat": "^2.6.1",
+    "gulp-postcss": "^7.0.0",
     "gulp-replace": "^0.5.4",
-    "gulp-sass": "^2.2.0",
-    "history": "^2.0.1",
-    "html-loader": "^0.4.0",
-    "html-webpack-plugin": "^2.7.1",
-    "http-proxy": "^1.12.0",
-    "json-loader": "^0.5.4",
-    "less": "^2.5.3",
-    "less-loader": "^2.2.1",
-    "node-sass": "^3.4.2",
-    "opn": "^4.0.0",
-    "postcss-loader": "^0.8.0",
-    "prop-types": "^15.5.8",
-    "react-router": "^2.0.0",
-    "react-transform-hmr": "^1.0.1",
-    "run-sequence": "^1.1.5",
-    "sass-loader": "^3.1.2",
-    "style-loader": "^0.13.0",
-    "task-dir": "^0.2.0",
-    "webpack": "^1.12.11",
-    "webpack-dev-middleware": "^1.2.0",
-    "webpack-hot-middleware": "^2.5.0",
-    "webpack-manifest-plugin": "^1.0.0",
-    "webpack-stream": "^3.0.1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/vn38minhtran/react-photoswipe.git"
-  },
-  "author": "Minh Tran",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/vn38minhtran/react-photoswipe/issues"
-  },
-  "homepage": "https://github.com/vn38minhtran/react-photoswipe"
+    "gulp-sass": "^3.1.0",
+    "html-webpack-plugin": "^2.28.0",
+    "http-proxy": "^1.16.2",
+    "node-sass": "^4.5.0",
+    "postcss-loader": "^2.0.5",
+    "prop-types": "^15.5.10",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-hot-loader": "^1.3.1",
+    "react-router-dom": "^4.1.1",
+    "run-sequence": "^1.2.2",
+    "sass-loader": "^6.0.1",
+    "style-loader": "^0.18.2",
+    "webpack": "^2.2.1",
+    "webpack-dev-middleware": "^1.6.1",
+    "webpack-hot-middleware": "^2.10.0",
+    "webpack-stream": "^3.2.0"
+  }
 }
