@@ -1,7 +1,5 @@
 import PhotoSwipe from './PhotoSwipe.js';
 import PhotoSwipeGallery from './PhotoSwipeGallery.js';
 
-module.exports = {
-  PhotoSwipe: PhotoSwipe,
-  PhotoSwipeGallery: PhotoSwipeGallery
-};
+export default PhotoSwipe;
+export { PhotoSwipe, PhotoSwipeGallery };
