@@ -1,19 +1,18 @@
 import url from 'url';
 import express from 'express';
 import webpack from 'webpack';
-import webpackConfig from './webpack.config';
 import webpackDevMiddleware from 'webpack-dev-middleware';
 import webpackHotMiddleware from 'webpack-hot-middleware';
 import fs from 'fs';
 import path from 'path';
 import http from 'http';
-import https from 'https'
-import opn from 'opn';
+import https from 'https';
 import httpProxy from 'http-proxy';
+import webpackConfig from './webpack.config.babel';
 
-const devURL = 'http://localhost:3000';
+const devURL = process.env.npm_package_config_devURL || 'http://localhost:3000';
 const urlParts = url.parse(devURL);
-const proxyOptions = [];
+const proxyOptions = process.env.npm_package_config_proxy || [];
 
 const proxy = httpProxy.createProxyServer({
   changeOrigin: true,
@@ -48,9 +47,9 @@ app.use(webpackHotMiddleware(compiler));
 
 app.use('/assets', express.static(path.join(__dirname, 'app/assets')));
 
-proxyOptions.forEach(option => {
+proxyOptions.forEach((option) => {
   app.all(option.path, (req, res) => {
-    proxy.web(req, res, option, err => {
+    proxy.web(req, res, option, (err) => {
       console.log(err.message);
       res.statusCode = 502;
       res.end();
@@ -59,14 +58,15 @@ proxyOptions.forEach(option => {
 });
 
 app.get('*', (req, res, next) => {
-  let filename = path.join(compiler.outputPath, 'index.html');
+  const filename = path.join(compiler.outputPath, 'index.html');
   compiler.outputFileSystem.readFile(filename, (error, result) => {
     if (error) {
-      return next(error);
+      next(error);
+    } else {
+      res.set('content-type', 'text/html');
+      res.send(result);
+      res.end();
     }
-    res.set('content-type', 'text/html');
-    res.send(result);
-    res.end();
   });
 });
 
@@ -79,6 +79,5 @@ if (urlParts.protocol === 'https:') {
 }
 
 server.listen(urlParts.port, () => {
-  console.log('Listening at ' + devURL);
-  opn(devURL);
+  console.log(`Listening at ${devURL}`);
 });
