@@ -1,87 +1,88 @@
 'use strict';
 
-Object.defineProperty(exports, '__esModule', {
+Object.defineProperty(exports, "__esModule", {
   value: true
 });
 
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
-var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
-
-var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_again) { var object = _x, property = _x2, receiver = _x3; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x = parent; _x2 = property; _x3 = receiver; _again = true; desc = parent = undefined; continue _function; } } else if ('value' in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
-
-function _objectWithoutProperties(obj, keys) { var target = {}; for (var i in obj) { if (keys.indexOf(i) >= 0) continue; if (!Object.prototype.hasOwnProperty.call(obj, i)) continue; target[i] = obj[i]; } return target; }
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
-
-function _inherits(subClass, superClass) { if (typeof superClass !== 'function' && superClass !== null) { throw new TypeError('Super expression must either be null or a function, not ' + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
 
 var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
-var _reactDom = require('react-dom');
+var _propTypes = require('prop-types');
 
-var _reactDom2 = _interopRequireDefault(_reactDom);
-
-var _PhotoSwipeJs = require('./PhotoSwipe.js');
-
-var _PhotoSwipeJs2 = _interopRequireDefault(_PhotoSwipeJs);
-
-var _lodashPick = require('lodash.pick');
-
-var _lodashPick2 = _interopRequireDefault(_lodashPick);
-
-var _events = require('./events');
-
-var _events2 = _interopRequireDefault(_events);
+var _propTypes2 = _interopRequireDefault(_propTypes);
 
 var _classnames = require('classnames');
 
 var _classnames2 = _interopRequireDefault(_classnames);
 
-var PhotoSwipeGallery = (function (_React$Component) {
+var _lodash = require('lodash.pick');
+
+var _lodash2 = _interopRequireDefault(_lodash);
+
+var _PhotoSwipe = require('./PhotoSwipe.js');
+
+var _PhotoSwipe2 = _interopRequireDefault(_PhotoSwipe);
+
+var _events = require('./events');
+
+var _events2 = _interopRequireDefault(_events);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+function _objectWithoutProperties(obj, keys) { var target = {}; for (var i in obj) { if (keys.indexOf(i) >= 0) continue; if (!Object.prototype.hasOwnProperty.call(obj, i)) continue; target[i] = obj[i]; } return target; }
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
+
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+
+var PhotoSwipeGallery = function (_React$Component) {
   _inherits(PhotoSwipeGallery, _React$Component);
 
   function PhotoSwipeGallery() {
-    var _this = this;
+    var _ref;
+
+    var _temp, _this, _ret;
 
     _classCallCheck(this, PhotoSwipeGallery);
 
-    _get(Object.getPrototypeOf(PhotoSwipeGallery.prototype), 'constructor', this).apply(this, arguments);
+    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
+      args[_key] = arguments[_key];
+    }
 
-    this.state = {
+    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_ref = PhotoSwipeGallery.__proto__ || Object.getPrototypeOf(PhotoSwipeGallery)).call.apply(_ref, [this].concat(args))), _this), _this.state = {
       isOpen: false,
-      options: this.props.options
-    };
-
-    this.showPhotoSwipe = function (itemIndex) {
+      options: _this.props.options
+    }, _this.showPhotoSwipe = function (itemIndex) {
       return function (e) {
         e.preventDefault();
-        var options = _this.state.options;
-
-        options.index = itemIndex;
-        options.getThumbBoundsFn = options.getThumbBoundsFn || function (index) {
-          var thumbnail = _reactDom2['default'].findDOMNode(_this.refs['thumbnail' + index]);
+        var getThumbBoundsFn = function getThumbBoundsFn(index) {
+          var thumbnail = _this.thumbnails[index];
           var img = thumbnail.getElementsByTagName('img')[0];
           var pageYScroll = window.pageYOffset || document.documentElement.scrollTop;
           var rect = img.getBoundingClientRect();
           return { x: rect.left, y: rect.top + pageYScroll, w: rect.width };
         };
+        var options = _this.state.options;
+
+        options.index = itemIndex;
+        options.getThumbBoundsFn = options.getThumbBoundsFn || getThumbBoundsFn;
         _this.setState({
           isOpen: true,
           options: options
         });
       };
-    };
-
-    this.handleClose = function () {
+    }, _this.handleClose = function () {
       _this.setState({
         isOpen: false
       });
-    };
+    }, _temp), _possibleConstructorReturn(_this, _ret);
   }
 
   _createClass(PhotoSwipeGallery, [{
@@ -89,66 +90,69 @@ var PhotoSwipeGallery = (function (_React$Component) {
     value: function render() {
       var _this2 = this;
 
-      var _props = this.props;
-      var id = _props.id;
-      var className = _props.className;
-      var items = _props.items;
-      var thumbnailContent = _props.thumbnailContent;
+      var _props = this.props,
+          id = _props.id,
+          items = _props.items,
+          thumbnailContent = _props.thumbnailContent,
+          other = _objectWithoutProperties(_props, ['id', 'items', 'thumbnailContent']);
 
-      var other = _objectWithoutProperties(_props, ['id', 'className', 'items', 'thumbnailContent']);
+      var className = this.props.className;
 
-      className = (0, _classnames2['default'])(['pswp-gallery', className]).trim();
-      var eventProps = (0, _lodashPick2['default'])(other, _events2['default']);
-      var _state = this.state;
-      var isOpen = _state.isOpen;
-      var options = _state.options;
+      className = (0, _classnames2.default)(['pswp-gallery', className]).trim();
+      var eventProps = (0, _lodash2.default)(other, _events2.default);
+      var _state = this.state,
+          isOpen = _state.isOpen,
+          options = _state.options;
 
-      return _react2['default'].createElement(
+      return _react2.default.createElement(
         'div',
         { id: id, className: className },
-        _react2['default'].createElement(
+        _react2.default.createElement(
           'div',
           { className: 'pswp-thumbnails' },
           items.map(function (item, index) {
-            return _react2['default'].createElement(
+            return _react2.default.createElement(
               'div',
-              { key: index, ref: 'thumbnail' + index,
+              {
+                key: index,
+                ref: function ref(node) {
+                  _this2.thumbnails = _this2.thumbnails || [];
+                  _this2.thumbnails[index] = node;
+                },
                 className: 'pswp-thumbnail',
-                onClick: _this2.showPhotoSwipe(index) },
+                onClick: _this2.showPhotoSwipe(index)
+              },
               thumbnailContent(item)
             );
           })
         ),
-        _react2['default'].createElement(_PhotoSwipeJs2['default'], _extends({}, eventProps, {
+        _react2.default.createElement(_PhotoSwipe2.default, _extends({}, eventProps, {
           isOpen: isOpen,
           items: items,
           options: options,
-          onClose: this.handleClose }))
+          onClose: this.handleClose
+        }))
       );
     }
-  }], [{
-    key: 'propTypes',
-    value: {
-      items: _react2['default'].PropTypes.array.isRequired,
-      options: _react2['default'].PropTypes.object,
-      thumbnailContent: _react2['default'].PropTypes.func,
-      id: _react2['default'].PropTypes.string,
-      className: _react2['default'].PropTypes.string
-    },
-    enumerable: true
-  }, {
-    key: 'defaultProps',
-    value: {
-      options: {},
-      thumbnailContent: function thumbnailContent(item) {
-        return _react2['default'].createElement('img', { src: item.src, width: '100', height: '100' });
-      }
-    },
-    enumerable: true
   }]);
 
   return PhotoSwipeGallery;
-})(_react2['default'].Component);
+}(_react2.default.Component);
 
-exports['default'] = PhotoSwipeGallery;
+PhotoSwipeGallery.propTypes = {
+  items: _propTypes2.default.array.isRequired,
+  options: _propTypes2.default.object,
+  thumbnailContent: _propTypes2.default.func,
+  id: _propTypes2.default.string,
+  className: _propTypes2.default.string
+};
+PhotoSwipeGallery.defaultProps = {
+  options: {},
+  thumbnailContent: function thumbnailContent(item) {
+    return _react2.default.createElement('img', { src: item.src, width: '100', height: '100', alt: '' });
+  },
+  id: '',
+  className: ''
+};
+exports.default = PhotoSwipeGallery;
 module.exports = exports['default'];
