@@ -1,16 +1,20 @@
 'use strict';
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.PhotoSwipeGallery = exports.PhotoSwipe = undefined;
 
-var _PhotoSwipeJs = require('./PhotoSwipe.js');
+var _PhotoSwipe = require('./PhotoSwipe.js');
 
-var _PhotoSwipeJs2 = _interopRequireDefault(_PhotoSwipeJs);
+var _PhotoSwipe2 = _interopRequireDefault(_PhotoSwipe);
 
-var _PhotoSwipeGalleryJs = require('./PhotoSwipeGallery.js');
+var _PhotoSwipeGallery = require('./PhotoSwipeGallery.js');
 
-var _PhotoSwipeGalleryJs2 = _interopRequireDefault(_PhotoSwipeGalleryJs);
+var _PhotoSwipeGallery2 = _interopRequireDefault(_PhotoSwipeGallery);
 
-module.exports = {
-  PhotoSwipe: _PhotoSwipeJs2['default'],
-  PhotoSwipeGallery: _PhotoSwipeGalleryJs2['default']
-};
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+exports.default = _PhotoSwipe2.default;
+exports.PhotoSwipe = _PhotoSwipe2.default;
+exports.PhotoSwipeGallery = _PhotoSwipeGallery2.default;
