@@ -1,10 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import ReactDOM from 'react-dom';
-import PhotoSwipe from './PhotoSwipe.js';
-import pick from 'lodash.pick';
-import events from './events';
 import classnames from 'classnames';
+import pick from 'lodash.pick';
+import PhotoSwipe from './PhotoSwipe.js';
+import events from './events';
 
 class PhotoSwipeGallery extends React.Component {
   static propTypes = {
@@ -17,11 +16,11 @@ class PhotoSwipeGallery extends React.Component {
 
   static defaultProps = {
     options: {},
-    thumbnailContent: (item) => {
-      return (
-        <img src={item.src} width="100" height="100"/>
-      );
-    }
+    thumbnailContent: item => (
+      <img src={item.src} width="100" height="100" alt=""/>
+    ),
+    id: '',
+    className: ''
   };
 
   state = {
@@ -29,23 +28,22 @@ class PhotoSwipeGallery extends React.Component {
     options: this.props.options
   };
 
-  showPhotoSwipe = (itemIndex) => {
-    return (e) => {
-      e.preventDefault();
-      let {options} = this.state;
-      options.index = itemIndex;
-      options.getThumbBoundsFn = options.getThumbBoundsFn || ((index) => {
-        let thumbnail = ReactDOM.findDOMNode(this.refs['thumbnail' + index]);
-        let img = thumbnail.getElementsByTagName('img')[0];
-        let pageYScroll = window.pageYOffset || document.documentElement.scrollTop;
-        let rect = img.getBoundingClientRect();
-        return {x: rect.left, y: rect.top + pageYScroll, w: rect.width};
-      });
-      this.setState({
-        isOpen: true,
-        options: options
-      });
-    };
+  showPhotoSwipe = itemIndex => (e) => {
+    e.preventDefault();
+    const getThumbBoundsFn = ((index) => {
+      const thumbnail = this.thumbnails[index];
+      const img = thumbnail.getElementsByTagName('img')[0];
+      const pageYScroll = window.pageYOffset || document.documentElement.scrollTop;
+      const rect = img.getBoundingClientRect();
+      return { x: rect.left, y: rect.top + pageYScroll, w: rect.width };
+    });
+    const { options } = this.state;
+    options.index = itemIndex;
+    options.getThumbBoundsFn = options.getThumbBoundsFn || getThumbBoundsFn;
+    this.setState({
+      isOpen: true,
+      options
+    });
   };
 
   handleClose = () => {
@@ -55,28 +53,35 @@ class PhotoSwipeGallery extends React.Component {
   };
 
   render() {
-    let {id, className, items, thumbnailContent, ...other} = this.props;
+    const { id, items, thumbnailContent, ...other } = this.props;
+    let { className } = this.props;
     className = classnames(['pswp-gallery', className]).trim();
-    let eventProps = pick(other, events);
-    let {isOpen, options} = this.state;
+    const eventProps = pick(other, events);
+    const { isOpen, options } = this.state;
     return (
       <div id={id} className={className}>
         <div className="pswp-thumbnails">
-          {items.map((item, index) => {
-            return (
-              <div key={index} ref={'thumbnail' + index}
-                className="pswp-thumbnail"
-                onClick={this.showPhotoSwipe(index)}>
-                {thumbnailContent(item)}
-              </div>
-            );
-          })}
+          {items.map((item, index) => (
+            <div
+              key={index}
+              ref={(node) => {
+                this.thumbnails = this.thumbnails || [];
+                this.thumbnails[index] = node;
+              }}
+              className="pswp-thumbnail"
+              onClick={this.showPhotoSwipe(index)}
+            >
+              {thumbnailContent(item)}
+            </div>
+          ))}
         </div>
-        <PhotoSwipe {...eventProps}
+        <PhotoSwipe
+          {...eventProps}
           isOpen={isOpen}
           items={items}
           options={options}
-          onClose={this.handleClose}/>
+          onClose={this.handleClose}
+        />
       </div>
     );
   }
