@@ -1,14 +1,13 @@
 import webpack from 'webpack';
-import pkg from './package.json';
+import path from 'path';
 import camelCase from 'camelcase';
+import pkg from './package.json';
 
-const capitalizeFirstLetter = (string) => {
-  return string.charAt(0).toUpperCase() + string.slice(1);
-};
+const capitalizeFirstLetter = string => string.charAt(0).toUpperCase() + string.slice(1);
 
 const webpackConfig = {
   output: {
-    filename: pkg.name + '.js',
+    filename: `${pkg.name}.js`,
     library: capitalizeFirstLetter(camelCase(pkg.name)),
     libraryTarget: 'umd'
   },
@@ -24,20 +23,39 @@ const webpackConfig = {
       commonjs: 'react-dom',
       commonjs2: 'react-dom',
       amd: 'react-dom'
+    },
+    'prop-types': {
+      root: 'PropTypes',
+      commonjs: 'prop-types',
+      commonjs2: 'prop-types',
+      amd: 'prop-types'
     }
   },
   module: {
-    loaders: [
+    rules: [
       {
         test: /\.(js|jsx)$/,
-        exclude: /(node_modules)/,
+        exclude: /node_modules/,
+        use: {
+          loader: 'eslint-loader',
+          options: {
+            configFile: path.join(__dirname, '.eslintrc'),
+            failOnError: true,
+            emitError: true
+          }
+        },
+        enforce: 'pre'
+      },
+      {
+        test: /\.(js|jsx)$/,
+        exclude: /node_modules/,
         loader: 'babel-loader'
       }
     ]
   },
   resolve: {
-    modulesDirectories: ['node_modules', 'bower_components'],
-    extensions: ['', '.jsx', '.js']
+    modules: ['node_modules'],
+    extensions: ['.jsx', '.js']
   },
   plugins: [
     new webpack.DefinePlugin({
@@ -53,8 +71,7 @@ const webpackConfig = {
       output: {
         comments: false
       }
-    }),
-    new webpack.optimize.DedupePlugin()
+    })
   ]
 };
 
